pyproject.toml
[tool.poetry]
name = "document_recognition"
version = "0.1.0"
description = ""
authors = ["Fom123 <gamemode1.459@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-vision = "2.7.3"
opencv-python = "4.5.4.60"
numpy = "^1.23.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
